duck-typing + len condition
