+++ force_nevergrad/engine/nevergrad_optimizers.py
@@ -10,131 +10,16 @@ from traits.api import (
 
 from force_bdss.api import (
     PositiveInt,
-    FixedMCOParameter,
-    RangedMCOParameter,
-    RangedVectorMCOParameter,
-    ListedMCOParameter,
-    CategoricalMCOParameter,
     KPISpecification,
     IOptimizer
 )
 
+from .parameter_translation import (
+    translate_mco_to_ng,
+    translate_ng_to_mco
+)
 import nevergrad as ng
 from nevergrad.functions import MultiobjectiveFunction
-
-
-def translate_mco_to_ng(params):
-    r""" Translate from an MCO parameter specification
-    to a Nevergrad parameter specification (Instrumentation).
-
-    Parameters
-    ----------
-    params: list of MCOParameter
-        The MCO parameter specification.
-
-    Return
-    ------
-    Instrumentation
-        Nevergrad instrumentation object.
-
-    Notes
-    -----
-    The Instrumentation object can be created from a list and/or
-    a dict of nevergrad parameter types. The keys of the dict are the
-    names of the parameters.
-    These list/dict is passed to the Instrumentation __init__ as
-    \*vargs/\*kwargs and they set the attributes called args/kwargs.
-    The result of a nevergrad optimization is itself a Instrumentation object.
-
-    As MCO objective functions (_score(), etc) take their parameters as a
-    single list argument, rather than kwargs (i.e. parameters are indexed
-    by position) we must create the Instrumentation by \*vargs. Also against
-    using \*kwargs, is that MCO parameter names may not be unique (there is
-    nothing to enforce this).
-
-    We could do with making many of the paramterization arguments attributes
-    of the optimizer.
-    """
-
-    instru = []
-    for p in params:
-
-        ng_param = None
-        if isinstance(p, FixedMCOParameter):
-            ng_param = ng.p.Constant(
-                value=p.value,
-            )
-        elif isinstance(p, RangedVectorMCOParameter):
-            ng_param = ng.p.Array(
-                init=np.array(p.initial_value),
-                mutable_sigma=True
-            )
-            ng_param.set_bounds(
-                lower=np.array(p.lower_bound),
-                upper=np.array(p.upper_bound),
-                method="arctan"
-            )
-        elif isinstance(p, RangedMCOParameter):
-            ng_param = ng.p.Scalar(
-                init=p.initial_value,
-                lower=p.lower_bound,
-                upper=p.upper_bound,
-                mutable_sigma=False
-            )
-        elif isinstance(p, ListedMCOParameter):
-            ng_param = ng.p.TransitionChoice(
-                choices=p.levels,
-                transitions=[1.0, 1.0]
-            )
-        elif isinstance(p, CategoricalMCOParameter):
-            ng_param = ng.p.Choice(
-                choices=p.categories,
-                deterministic=False
-            )
-        else:
-            # duck-typing for non-standard??? quack, quack
-            ng_param = ng.p.Constant(
-                value="null"
-            )
-
-        # add parameter to list
-        instru.append(ng_param)
-
-    # create Instrumentation object with *vargs
-    return ng.p.Instrumentation(*instru)
-
-
-def translate_ng_to_mco(ng_params):
-    """ Translate a list of nevergrad parameter values
-    to a list of MCO parameter values.
-
-    Parameters
-    ----------
-    ng_params: list of Any (but usually float, ndarray or string)
-        Parameter values in the nevergrad form
-
-    Return
-    ------
-    mco_values: list of Any (but usually float, list or string)
-        Parameter values in the MCO form
-
-    Notes
-    -----
-    These are mostly the same, except for RangedVectorMCOParameter to
-    ng.p.Array conversion: the value of the former is a list, whereas
-    the value of the latter is a numpy array
-    """
-
-    mco_values = []
-    for p in ng_params:
-        if isinstance(p, np.ndarray):
-            mco_values.append(p.tolist())
-        else:
-            mco_values.append(p)
-        # ...what about any non-standard MCOParameter types?
-        # (see translate_mco_to_ng(), above)
-
-    return mco_values
 
 
 def nevergrad_function(*ng_params,
@@ -184,9 +69,9 @@ def nevergrad_function(*ng_params,
     objective = function(mco_params)
 
     # negate any objectives that should be maximised
-    if minimize_objectives:
+    if len(minimize_objectives) == len(objective):
         for i, do_minimize in enumerate(minimize_objectives):
-            if (not do_minimize) and (i < len(objective)):
+            if not do_minimize:
                 objective[i] = -objective[i]
 
     # if the objective needs to be scalar but it is not,
@@ -293,19 +178,19 @@ class NevergradMultiOptimizer(HasStrictTraits):
             The list of parameter values for a single member
             of the Pareto set.
         """
-
+        print('c0, func = ', func.__name__)
         # Create instrumentation.
         instrumentation = translate_mco_to_ng(params)
-
+        print('c1')
         # Create optimizer.
         optimizer = ng.optimizers.registry[self.algorithms](
             parametrization=instrumentation,
             budget=self.budget
         )
-
+        print('c2')
         # Minimization/maximization specification
         minimize_objectives = ['MINI' in k.objective for k in self.kpis]
-
+        print('c3, n objectives = ', len(minimize_objectives))
         # Create a multi-objective nevergrad function from
         # the MCO function.
         ng_func = partial(nevergrad_function,
@@ -313,16 +198,16 @@ class NevergradMultiOptimizer(HasStrictTraits):
                           is_scalar=False,
                           minimize_objectives=minimize_objectives
                           )
-
+        print('c4')
         # Create a MultiobjectiveFunction object from that.
         ob_func = MultiobjectiveFunction(
             multiobjective_function=ng_func,
             upper_bounds=np.array([k.scale_factor for k in self.kpis])
         )
-
+        print('c5')
         # Optimize. Ignore the return.
         optimizer.minimize(ob_func)
-
+        print('c6, length of pareto = ', len(ob_func.pareto_front()))
         # yield a member of the Pareto set.
         # x is a tuple - ((<vargs parameters>), {<kwargs parameters>})
         # return the vargs, translated into mco.
