+++ force_nevergrad/mco/ng_mco.py
@@ -43,9 +43,10 @@ class NevergradMCO(BaseMCO):
         screen_handler = logging.StreamHandler(stream=sys.stdout)
         screen_handler.setFormatter(formatter)
         log.addHandler(screen_handler)
-
+        print('a0')
         for index, (optimal_point, optimal_kpis) \
                 in enumerate(optimizer.optimize()):
+            print('a1')
             # When there is new data, this operation informs the system that
             # new data has been received. It must be a dictionary as given.
             log.info("Doing  MCO run # {}".format(index))
